:hammer: implementing proof of work consensus...
